==> workflow/00-create_d2m_s2m_profile.Rmd <==
---
title: "00-create_d2m_s2m_profile"
date: "Compiled at `r format(Sys.time(), '%Y-%m-%d %H:%M:%S')`"
output: github_document
params:
  name: "00-create_d2m_s2m_profile"
---

```{r here, message=FALSE, include=FALSE}
here::i_am(paste0(params$name, ".Rmd"), uuid = "17d88d45-e612-4afa-a93a-c74025e9f801")
```

The purpose of this document is ...

```{r packages, message=FALSE, warning=FALSE}
library("conflicted")
library(here)
library(readxl)
library(tidyverse)

prj_path <- normalizePath("..")
fpath <- here(prj_path, "data/disease.xlsx")

# for sm2miRNA profile normalization
deepsmirud_threshold <- 0.5
```

```{r directories, include=FALSE}
# create or *empty* the target directory, used to write this file's data: 
projthis::proj_create_dir_target(params$name, clean = F)

# function to get path to target directory: path_target("sample.csv")
path_target <- projthis::proj_path_target(params$name)

# function to get path to previous data: path_source("00-import", "sample.csv")
path_source <- projthis::proj_path_source(params$name)
```

## Tasks

The first task is ...

-   Non-human miRNA were removed from DeepsmirUD predictions. 6841

### Disease-miRNA regulation profile

The disease-miRNA regulation profile is defined as a miRNA-disease matrix converted from wide to long format, and stored in `df_miRNA2disease` dataframe. Values in the matrix is either "up" or "down" indicate the miRNA is up or down-regulated in the disease. The regulation profile was obtained from [miRCancer](http://mircancer.ecu.edu/) database.

```{r Import disease-miRNA regulation profile, message=FALSE, warning=FALSE}
df_disease <- read_excel(fpath, sheet = "mircancer")
disease_names <- unique(df_disease$disease_name)

# disease-miRNA profile
profile_d2m <- df_disease %>% 
  mutate(miRNA_reg_profile = Profile, .keep = "unused") %>% 
  dplyr::select(c(disease_name, mirna_baseid, miRNA_reg_profile))
```

There are `r length(unique(df_disease$mirna_baseid))` miRNAs involved in `r length(disease_names)` diseases (connectivity: `r 100 * nrow(profile_d2m) / (length(unique(df_disease$mirna_baseid)) * length(disease_names))`%).

### SM-miRNA regulation profile

The regulation profile was predicted using DeepsmirUD.

```{r Import compound-miRNA regulation profile, message=FALSE, warning=FALSE}
df_deepsmirud <- read_excel(fpath, sheet = "deepsmirud")

# remove non-human miRNA
df_deepsmirud_human <- df_deepsmirud %>% 
  dplyr::filter(is_exp == "Y") %>% 
  dplyr::filter(str_detect(mir_name, "^hsa") | str_detect(mir_name, "^mi"))

# remove SM-miRNA pair redundancy by `mirna_baseid`
# TODO: use more proper way to reduce redundancy
df_deepsmirud_human_nr <- df_deepsmirud_human %>% 
  dplyr::select(c(mirna_baseid, sm_name, DeepsmirUD_heatmap)) %>% 
  distinct(mirna_baseid, sm_name, .keep_all = T)

# SM-miRNA profile in wide format
profile_s2m_raw <- df_deepsmirud_human_nr %>% 
  pivot_wider(names_from = sm_name, values_from = DeepsmirUD_heatmap) %>% 
  column_to_rownames("mirna_baseid")


# --------- normalize and format SM-miRNA profile matrix -----------
profile_s2m <- profile_s2m_raw - deepsmirud_threshold
profile_s2m[is.na(profile_s2m)] <- 0

# Change SM name so that they are distinct and satisfy DF column name requirement.
colnames_old <- colnames(profile_s2m)
colnames(profile_s2m) <- make.names(colnames(profile_s2m), unique = T)
colnames_new <- colnames(profile_s2m)
df_smname_mapping <- data.frame(colnames_new = colnames_new,
                                colnames_old = colnames_old)

save(profile_d2m, profile_s2m, df_smname_mapping, disease_names, file = path_target("profiles_deepsmirud_ref.Rdata"))
nrow(profile_s2m) / (nrow(profile_s2m) * ncol(profile_s2m))
```

|              | Original                                       | Human                                                | Distinct              |
|--------------|------------------------------------------------|------------------------------------------------------|-----------------------|
| Mirbase      | `r length(unique(df_deepsmirud$Mirbase))`      | `r length(unique(df_deepsmirud_human$Mirbase))`      |                       |
| mirna_baseid | `r length(unique(df_deepsmirud$mirna_baseid))` | `r length(unique(df_deepsmirud_human$mirna_baseid))` | `r nrow(profile_s2m)` |

There are `r nrow(profile_s2m)` miRNAs involved in `r ncol(profile_s2m)` small compounds (connectivity: `r nrow(profile_s2m) / (nrow(profile_s2m) * ncol(profile_s2m))`%).

```{r}
df_deepsmirud <- read_excel(fpath, sheet = "deepsmirud")

# remove non-human miRNA
df_deepsmirud_human <- df_deepsmirud %>% 
  dplyr::filter(is_exp == "Y") %>% 
  dplyr::filter(str_detect(mir_name, "^hsa") | str_detect(mir_name, "^mi"))

# remove SM-miRNA pair redundancy by `mirna_baseid`
# TODO: use more proper way to reduce redundancy
df_deepsmirud_human_nr <- df_deepsmirud_human %>% 
  dplyr::select(c(mirna_baseid, sm_name, exp)) %>% 
  distinct(mirna_baseid, sm_name, .keep_all = T)

# SM-miRNA profile in wide format
profile_s2m_raw <- df_deepsmirud_human_nr %>% 
  pivot_wider(names_from = sm_name, values_from = exp) %>% 
  column_to_rownames("mirna_baseid")


# --------- normalize and format SM-miRNA profile matrix -----------
profile_s2m <- profile_s2m_raw - deepsmirud_threshold
profile_s2m[is.na(profile_s2m)] <- 0

# Change SM name so that they are distinct and satisfy DF column name requirement.
colnames_old <- colnames(profile_s2m)
colnames(profile_s2m) <- make.names(colnames(profile_s2m), unique = T)
colnames_new <- colnames(profile_s2m)
df_smname_mapping <- data.frame(colnames_new = colnames_new,
                                colnames_old = colnames_old)

save(profile_d2m, profile_s2m, df_smname_mapping, disease_names, file = path_target("profiles_exp_ref.Rdata"))
nrow(profile_s2m) / (nrow(profile_s2m) * ncol(profile_s2m))
```

```{r}
df_deepsmirud <- read_excel(fpath, sheet = "deepsmirud")

# remove non-human miRNA
df_deepsmirud_human <- df_deepsmirud %>% 
  dplyr::filter(is_exp == "N") %>% 
  dplyr::filter(str_detect(mir_name, "^hsa") | str_detect(mir_name, "^mi"))

# remove SM-miRNA pair redundancy by `mirna_baseid`
# TODO: use more proper way to reduce redundancy
df_deepsmirud_human_nr <- df_deepsmirud_human %>% 
  dplyr::select(c(mirna_baseid, sm_name, DeepsmirUD_heatmap)) %>% 
  distinct(mirna_baseid, sm_name, .keep_all = T)

# SM-miRNA profile in wide format
profile_s2m_raw <- df_deepsmirud_human_nr %>% 
  pivot_wider(names_from = sm_name, values_from = DeepsmirUD_heatmap) %>% 
  column_to_rownames("mirna_baseid")


# --------- normalize and format SM-miRNA profile matrix -----------
profile_s2m <- profile_s2m_raw - deepsmirud_threshold
profile_s2m[is.na(profile_s2m)] <- 0

# Change SM name so that they are distinct and satisfy DF column name requirement.
colnames_old <- colnames(profile_s2m)
colnames(profile_s2m) <- make.names(colnames(profile_s2m), unique = T)
colnames_new <- colnames(profile_s2m)
df_smname_mapping <- data.frame(colnames_new = colnames_new,
                                colnames_old = colnames_old)

save(profile_d2m, profile_s2m, df_smname_mapping, disease_names, file = path_target("profiles_deepsmirud_novo.Rdata"))
nrow(profile_s2m) / (nrow(profile_s2m) * ncol(profile_s2m))
```

```{r}
df_deepsmirud <- read_excel(fpath, sheet = "deepsmirud")

# remove non-human miRNA
df_deepsmirud_human <- df_deepsmirud %>% 
  dplyr::filter(str_detect(mir_name, "^hsa") | str_detect(mir_name, "^mi"))

# remove SM-miRNA pair redundancy by `mirna_baseid`
# TODO: use more proper way to reduce redundancy
df_deepsmirud_human_nr <- df_deepsmirud_human %>% 
  dplyr::select(c(mirna_baseid, sm_name, DeepsmirUD_heatmap)) %>% 
  distinct(mirna_baseid, sm_name, .keep_all = T)

# SM-miRNA profile in wide format
profile_s2m_raw <- df_deepsmirud_human_nr %>% 
  pivot_wider(names_from = sm_name, values_from = DeepsmirUD_heatmap) %>% 
  column_to_rownames("mirna_baseid")


# --------- normalize and format SM-miRNA profile matrix -----------
# Because cmap use log-fold change information, we need to convert the raw DeepsmirUD to [negative, positive] by minus a threshold (default: 0.5)
profile_s2m <- profile_s2m_raw - deepsmirud_threshold
profile_s2m[is.na(profile_s2m)] <- 0
# row * col = miRNA * small_molecule (SM)
# Rank order of miRNA in each SM was ordered in the RCSM code: https://github.com/Jasonlinchina/RCSM/blob/8f1d614ffd400225566e93f16d95c824e8a1606b/R/KS.R#L42

# Change SM name so that they are distinct and satisfy DF column name requirement.
colnames_old <- colnames(profile_s2m)
colnames(profile_s2m) <- make.names(colnames(profile_s2m), unique = T)
colnames_new <- colnames(profile_s2m)
df_smname_mapping <- data.frame(colnames_new = colnames_new,
                                colnames_old = colnames_old)

save(profile_d2m, profile_s2m, df_smname_mapping, disease_names, file = path_target("profiles_deepsmirud.Rdata"))
nrow(profile_s2m) / (nrow(profile_s2m) * ncol(profile_s2m))
```

```{r Drug-cancer heatmap}
load(here(prj_path, "data/row_col_order.Rdata"))
cmap_g1 <- readRDS(here(prj_path, "data/main/GSEAweight1Score.rds"))
cmap_rst <- cmap_g1

ftitle <- "GSEAweight1"
minN_miRNA <- 5
total_miRNA <- 10
pval <- 0.05
cluster <- T
# row_order=F
# col_order=F
outdir = "interactive"


df <- data.frame(sm=names(profile_s2m))
for (disease in names(cmap_rst)) {
  profile_d2m_sub <- profile_d2m %>%
    dplyr::filter(disease_name == disease) %>%
    dplyr::select(c(mirna_baseid, miRNA_reg_profile)) %>%
    distinct() %>%
    group_by(mirna_baseid) %>%
    dplyr::filter(n()==1)    # remove miRNA that are both up and down-regulated in a disease

  # up-regulated miRNA
  qry_up <- profile_d2m_sub %>%
    dplyr::filter(miRNA_reg_profile=="up") %>%
    pull(mirna_baseid)

  # down-ragulated miRNA
  qry_down <- profile_d2m_sub %>%
    dplyr::filter(miRNA_reg_profile=="down") %>%
    pull(mirna_baseid)

  # print(disease)
  # print(dim(profile_d2m_sub))
  cmap_one_disease <- cmap_rst[[disease]]
  if (nrow(cmap_one_disease)>0) {
    # If there are no enough miRNA for calculating cmap score, set their cmap score to 0
    if (length(qry_up) < minN_miRNA | length(qry_down) < minN_miRNA) {
      if (nrow(profile_d2m_sub) < total_miRNA) {
        cmap_one_disease["Score"] = 0
        cmap_one_disease["pValue"] = 1
        cmap_one_disease["pAdjValue"] = 1
      }
    }

    # print(disease)
    # return(cmap_one_disease)
    cmap_disease <- cmap_one_disease %>%
      dplyr::filter(pAdjValue < pval) %>%
      data.table::setnames("Score", disease) %>%
      dplyr::select(disease) %>%
      rownames_to_column("sm")
    df <- df %>%
      left_join(cmap_disease, by = "sm")
  }
}

df2 <- df %>%
  mutate(colnames_new = sm, .keep = "unused") %>%
  left_join(df_smname_mapping, by = "colnames_new") %>%
  dplyr::select(-colnames_new) %>%
  column_to_rownames("colnames_old") %>%
  replace(is.na(.), 0) %>%
  t()


df2 <- df2/2
rg <- max(abs(df2))


heatmaply::heatmaply(df2[row_order, col_order],
                            colors = colorRampPalette(c("navy", "white", "red"))(100),
                            grid_gap = 0,
                            fontsize_row = 10,
                            fontsize_col = 10,
                            width = 1800,
                            height = 700,
                            Rowv = F,
                            Colv = F,
                            xlab = "Small molecule",
                            ylab = "Cancer",
                            dynamicTicks = T,
                            # cellnote_size = 4,
                            margin = c(10,10,10,0),
                            file = "drug-cancer-heatmap.png",
                            limits = c(-rg, rg)
                     )
```


## Files written

These files have been written to the target directory, `r paste0("data/", params$name)`:

```{r list-files-target}
projthis::proj_dir_info(path_target())
```
