==> workflow/01-format_cmap_score.Rmd <==
---
title: "01-format_cmap_score"
date: "Compiled at `r format(Sys.time(), '%Y-%m-%d %H:%M:%S')`"
output: github_document
zotero: true
params:
  name: "01-format_cmap_score" # change if you rename file
---

```{r here, message=FALSE}
here::i_am(paste0(params$name, ".Rmd"), uuid = "9fe50057-4d4c-47d3-96e3-5e0f783dddc3")
```

The purpose of this document is ...

```{r packages}
library("conflicted")
library(here)
library(tidyverse)
library(pheatmap)
library(patchwork)
library("grid")
library("gridExtra")
```

```{r directories}
# create or *empty* the target directory, used to write this file's data: 
projthis::proj_create_dir_target(params$name, clean = TRUE)

# function to get path to target directory: path_target("sample.csv")
path_target <- projthis::proj_path_target(params$name)

# function to get path to previous data: path_source("00-import", "sample.csv")
path_source <- projthis::proj_path_source(params$name)
```

## Tasks

The first task is ...

```{r load data}
prj_path <- normalizePath("..")

# outdir <- "dud"
# dpath <- here(prj_path, "data/dud")
# load(path_source("00-create_d2m_s2m_profile", "profiles_deepsmirud_ref.Rdata"))
# # 
# outdir <- "exp"
# dpath <- here(prj_path, "data/exp")
# load(path_source("00-create_d2m_s2m_profile", "profiles_exp_ref.Rdata"))

# main
dpath <- here(prj_path, "data/main")
load(path_source("00-create_d2m_s2m_profile", "profiles_deepsmirud.Rdata"))
```

```{r load functions}
plt_heatmap <- function(cmap_profile, profile_s2m, profile_d2m, df_smname_mapping, pval=0.05, ftitle="cmap", minN_miRNA=5, total_miRNA=10, cluster = T, row_order=F, col_order=F, outdir = "heatmap", interactive=F) {
  cmap_rst <- cmap_profile
  
  df <- data.frame(sm=names(profile_s2m))
  for (disease in names(cmap_rst)) {
    profile_d2m_sub <- profile_d2m %>%
      dplyr::filter(disease_name == disease) %>%
      dplyr::select(c(mirna_baseid, miRNA_reg_profile)) %>%
      distinct() %>%
      group_by(mirna_baseid) %>%
      dplyr::filter(n()==1)    # remove miRNA that are both up and down-regulated in a disease
    
    # up-regulated miRNA
    qry_up <- profile_d2m_sub %>%
      dplyr::filter(miRNA_reg_profile=="up") %>%
      pull(mirna_baseid)
  
    # down-ragulated miRNA
    qry_down <- profile_d2m_sub %>%
      dplyr::filter(miRNA_reg_profile=="down") %>%
      pull(mirna_baseid)
    
    # print(disease)
    # print(dim(profile_d2m_sub))
    cmap_one_disease <- cmap_rst[[disease]]
    if (nrow(cmap_one_disease)>0) {
      # If there are no enough miRNA for calculating cmap score, set their cmap score to 0
      if (length(qry_up) < minN_miRNA | length(qry_down) < minN_miRNA) {
        if (nrow(profile_d2m_sub) < total_miRNA) {
          cmap_one_disease["Score"] = 0
          cmap_one_disease["pValue"] = 1
          cmap_one_disease["pAdjValue"] = 1
        }
      }
      
      # print(disease)
      # return(cmap_one_disease)
      cmap_disease <- cmap_one_disease %>%
        dplyr::filter(pAdjValue < pval) %>%
        data.table::setnames("Score", disease) %>%
        dplyr::select(disease) %>%
        rownames_to_column("sm")
      df <- df %>%
        left_join(cmap_disease, by = "sm")
    }
  }

  df2 <- df %>%
    mutate(colnames_new = sm, .keep = "unused") %>%
    left_join(df_smname_mapping, by = "colnames_new") %>%
    dplyr::select(-colnames_new) %>%
    column_to_rownames("colnames_old") %>%
    replace(is.na(.), 0) %>%
    t()
  
  # RCSM doesn't calculate average of up and down-regulated features,
  # so here we divide RCSM calculated scores by 2 so that final scores in [-1, 1].
  df2 <- df2/2
  
  if (row_order != FALSE) {
    df2 <- df2[row_order,]
  }
  if (col_order != FALSE) {
    df2 <- df2[, col_order]
  }

  rg <- max(abs(df2))
  p <- pheatmap(df2, fontsize_row = 1.6, fontsize_col = 0.5, width = 19, height = 5,
                filename = paste0(outdir, "/disease2sm_heatmap_", ftitle, ".pdf"),
                main = paste0(ftitle, " Score ", nrow(df2), " * ", ncol(df2)),
                cluster_cols = cluster, cluster_rows = cluster,
                color=colorRampPalette(c("navy", "white", "red"))(100),
                breaks = seq(-rg, rg, length.out = 100))
  return(p)
}
```

```{r format cmap score G1, fig.height=5, fig.width=18}
cmap_ks <- readRDS(here(dpath, "KSScore.rds"))
cmap_zh <- readRDS(here(dpath, "ZhangScore.rds"))
cmap_g0 <- readRDS(here(dpath, "GSEAweight0Score.rds"))
cmap_g2 <- readRDS(here(dpath, "GSEAweight2Score.rds"))
# cmap_xs <- readRDS(here(dpath, "XSumScore.rds"))

cluster <- F
totalmiRNA <- 10
minNmiRNA <- 5
outdir <- paste0("main_", as.character(totalmiRNA), "_", as.character(minNmiRNA))
dir.create(outdir)
cmap_g1 <- readRDS(here(dpath, "GSEAweight1Score.rds"))
pval <- 0.05
fig_g1 <- plt_heatmap(cmap_g1, profile_s2m, profile_d2m, df_smname_mapping, pval, "GSEAweight1", minNmiRNA, totalmiRNA, T, F, F, outdir)
row_order <- fig_g1$tree_row$labels[fig_g1$tree_row$order]
col_order <- fig_g1$tree_col$labels[fig_g1$tree_col$order]

save(row_order, col_order, file = here(prj_path, "data/row_col_order.Rdata"))
# row_order <- F
# col_order <- F

fig_g2 <- plt_heatmap(cmap_g2, profile_s2m, profile_d2m, df_smname_mapping, pval, "GSEAweight2", minNmiRNA, totalmiRNA,  cluster, row_order, col_order, outdir)
fig_zh <- plt_heatmap(cmap_zh, profile_s2m, profile_d2m, df_smname_mapping, pval, "Zhang", minNmiRNA, totalmiRNA, cluster, row_order, col_order, outdir)
fig_ks <- plt_heatmap(cmap_ks, profile_s2m, profile_d2m, df_smname_mapping, pval, "KS", minNmiRNA, totalmiRNA, cluster, row_order, col_order, outdir)
fig_g0 <- plt_heatmap(cmap_g0, profile_s2m, profile_d2m, df_smname_mapping, pval, "GSEAweight0", minNmiRNA, totalmiRNA, cluster, row_order, col_order, outdir)
# fig_xs <- plt_heatmap(cmap_xs, profile_s2m, profile_d2m, df_smname_mapping, pval, "XSum", minNmiRNA, cluster, row_order, col_order, outdir)

fig_compare <- grid.arrange(grobs = list(fig_g1[[4]],
                                         fig_g2[[4]],
                                         fig_zh[[4]],
                                         # fig_xs[[4]],
                                         fig_ks[[4]],
                                         fig_g0[[4]]),
                            layout_matrix = matrix(1:5, ncol = 1))

ggsave(file=paste0(outdir, "/", outdir, "_disease2sm_heatmap.pdf"), fig_compare, width = 20, height = 25)
```

```{r most prevalent SMs, fig.height=5, fig.width=9}
# df2 from `data4heatmap.R`
a <- colSums(df2 < -0)

DT::datatable(data.frame(num_disease = sort(a[a > 12], decreasing=T)))

ctop <- sort(a[a>12], decreasing=T)
df_sel <- df2[, names(ctop)]

b <- rowSums(df_sel != 0)
rtop <- sort(b[b>3], decreasing=T)

pheatmap(df2[names(rtop), names(ctop)], cluster_rows = F, cluster_cols = F,
         color=colorRampPalette(c("navy", "white", "red"))(100),
         main = "Top10 small molecules relavent to most number of cancers",
         breaks = seq(-1, 1, length.out = 100))


rsturl <- paste0("https://www.google.com/search?q=", head(names(ctop), 10), "+%2Bcancer")
# lapply(rsturl, browseURL)
```

## Files written

These files have been written to the target directory, `r paste0("data/", params$name)`:

```{r list-files-target}
projthis::proj_dir_info(path_target())
```
